Create cdw-getFileSize__FINAL_2025-1-10.ps1
Fixed an issue with $gciDIR_sum.Value and I had to cast it as a [double] to convert the string into an integer value and this resolved $Error that was storing the soft errors of not converting properly. The $Error continues to log directories that the script does not have access to, but that is a valid error to log and the only other thing needing verified/testing is the remote execution of the script on another Windows system.